External processing patch
This patch adds a new binary, and a new container, using the external
processing filter:

https://www.envoyproxy.io/docs/envoy/latest/api-v3/extensions/filters/http/ext_proc/v3/ext_proc.proto#external-processing-filter

This replaces the Lua filter. This increases latency when the system is
not loaded, but greatly increases throughput at scale.

There are still some rough edges related to configuration reloading:

 * because auto reloading is enabled, the code is a bit more complex
   than it should be, because it is not possible to hold a reference to
   the configuration. As a result, a lot of data needs to be copied and
   kept, instead of being a simple reference to the configuration.

 * it is now possible to get concurrent requests asking for
   configuration reloads, which get queued and might choke the system.

It also modifies curielogger so that it can parse the new log format.

Signed-off-by: Simon Marechal <bartavelle@gmail.com>