New curie session request information
User visible changes:
 * new field in security policies: `session`. Format is identical to
   limit's `key`, but semantics are distinct. When missing, default to
   the empty list.
 * new `curiesession` entry in the logs field.

The session field is used to select information about the request. The
first matching entry is kept, subsequent entries are ignored. If the
session field was empty, it defaults to getting the request source IP.

The selected information is then hashed with the current content filter
profile masking key.

Signed-off-by: Simon Marechal <bartavelle@gmail.com>